auth-service/requirements.txt
fastapi
mysql-connector-python
passlib
itsdangerous
starlette
slowapi
python-dotenv
